requirements.txt
anthropic
openpyxl
Pillow
protobuf
streamlit
